requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
joblib==1.5.2
numpy==2.2.6
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
SQLAlchemy==2.0.43
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
